Fixes; useClickAway moved to common-hooks
