+++ packages/lab/src/dropdown-next/DropdownBase.tsx
@@ -13,7 +13,7 @@ import { useDropdownBase } from "./useDropdownBase";
 import { Portal } from "../portal";
 import { isDesktop, useWindow } from "../window";
 
-import "./Dropdown.css";
+import "./DropdownNext.css";
 
 // Any component may be passed as our trigger or popup component.
 // Define the common props that we will act on, if present,
